+++ app/components/cards/Server.vue
@@ -49,37 +49,6 @@
       </div>
     </div>
   </div>
-
-  <!-- <div v-if="variant === 1" class="flex flex-col h-full">
-      <div class="relative w-full h-[180px] overflow-hidden bg-gray-100 shrink-0">
-        <img :src="image" alt="Item Image" class="w-full h-full object-cover" />
-        <div class="absolute inset-0 bg-gradient-to-t from-tertiary via-white/10 to-transparent pointer-events-none"></div>
-      </div>
-      <div class="p-5 flex flex-col grow">
-        <div class="mb-4">
-          <h3 class="font-inter text-2xl font-bold text-black mb-2">{{ title }}</h3>
-          <p class="font-inter text-xs text-gray-500">
-            {{ description }}
-          </p>
-        </div>
-        <div class="w-full h-1 bg-gray-100 rounded-full mb-4">
-          <div class="h-full bg-gray-200 rounded-full"></div>
-        </div>
-        <div class="px-5 mb-5">
-          <div class="flex justify-between items-center">
-            <p class="font-inter text-xs font-bold text-gray-400 mb-0.5">Tersedia</p>
-            <p class="font-inter text-sm font-bold text-black">{{ stock }} Item</p>
-          </div>
-          <div class="flex justify-between items-center">
-            <p class="font-inter text-xs font-bold text-gray-400 mb-0.5">Mulai Dari</p>
-            <p class="font-inter text-lg font-bold text-primary">Rp{{ formatPrice(price) }}</p>
-          </div>
-        </div>
-        <ButtonsPrimary variant="solid" color="secondary">
-          Beli Sekarang
-        </ButtonsPrimary>
-      </div>
-    </div> -->
 </template>
 
 <script setup lang="ts">